chore: setup PollEditor.tsx for Bran7astic
