--- src/components/poll/edit/UploadFileBox.tsx
@@ -1,13 +1,0 @@
-import { Typography } from "@mui/material"
-import React from "react"
-
-export default function UploadFileBox() {
-  /* Note - do not remove fragment. */
-  return (
-    <React.Fragment>
-      <Typography>
-        <code>Upload File Box</code>
-      </Typography>
-    </React.Fragment>
-  )
-}
